chore(specifications): move to /packages
