+++ packages/administration/src/specification-mng/config-validator/config-error.model.ts
