+++ packages/administration/src/specification-mng/config-validator/config-validator.util.ts
