+++ packages/administration/src/specification-mng/config-validator/schema-spec.ts
