+++ packages/administration/src/specification-mng/config-validator/schema.ts
