+++ packages/administration/src/specification-mng/config-visualizer-dialog/config-visualizer-dialog.component.scss
