+++ packages/administration/src/specification-mng/entity-spec-mng/entity-spec-mng.component.html
