+++ packages/administration/src/specification-mng/index.ts
@@ -1,0 +1,2 @@
+export { SpecificationMngModule } from './specification-mng.module';
+export { SpecificationMngComponent } from './specification-mng.component';
