+++ packages/administration/src/specification-mng/specification-mng.component.html
